restructured stories by type
