+++ .storybook/src/layers/canvas/callout-canvas.layer.stories.ts
@@ -1,10 +1,9 @@
 import { scaleLinear } from 'd3-scale';
 
-import { CalloutCanvasLayer } from '../../../src/layers';
-import { ZoomPanHandler } from '../../../src/control/ZoomPanHandler';
-import { Annotation, OnUpdateEvent } from '../../../src/interfaces';
+import { CalloutCanvasLayer } from '../../../../src/layers';
+import { Annotation } from '../../../../src/interfaces';
 
-import { createLayerContainer, createRootContainer, createFPSLabel } from '../utils';
+import { createLayerContainer, createRootContainer } from '../../utils';
 
 const annotations: Annotation[] = [
   {
@@ -26,15 +25,6 @@ const annotations: Annotation[] = [
     data: [460, 110],
   },
   {
-    title: 'Balder Fm. Top',
-    md: 1234.3,
-    tvd: 1234,
-    mdUnit: 'm',
-    depthReferencePoint: 'RKB',
-    group: 'strat-picks',
-    data: [460, 110],
-  },
-  {
     title: 'Odin Fm. Top',
     md: 1234.3,
     tvd: 1234,
@@ -42,15 +32,6 @@ const annotations: Annotation[] = [
     depthReferencePoint: 'RKB',
     group: 'strat-picks',
     data: [350, 60],
-  },
-  {
-    title: 'Loke Fm. 2.1 Top',
-    md: 1234,
-    tvd: 1234,
-    mdUnit: 'm',
-    depthReferencePoint: 'RKB',
-    group: 'strat-picks',
-    data: [40, 70],
   },
   {
     title: 'Loke Fm. 2.1 Top',
@@ -88,11 +69,8 @@ const margin = {
   left: 0,
 };
 
-const xbounds: [number, number] = [0, 500];
-const ybounds: [number, number] = [0, 500];
-
-const xRange = 500;
-const yRange = 500;
+const xbounds = [0, 500];
+const ybounds = [0, 500];
 
 const xscale = scaleLinear()
   .domain(xbounds)
@@ -104,31 +82,15 @@ const yscale = scaleLinear()
 const width = 500;
 const height = 500;
 
-export const CalloutCanvasWithZoom = () => {
+export const CalloutCanvas = () => {
   const root = createRootContainer(width);
   const container = createLayerContainer(width, height);
 
   const layer = new CalloutCanvasLayer('callout', { order: 1 });
   layer.onMount(createEventObj(container));
-  layer.onUpdate(createEventObj(container));
-
-  const zoomHandler = new ZoomPanHandler(container, (event: OnUpdateEvent) => {
-    layer.onRescale({
-      ...event,
-      data: annotations,
-      annotations,
-      isLeftToRight: true,
-      margin,
-      scale: 0,
-    });
-  });
-
-  zoomHandler.setBounds(xbounds, ybounds);
-  zoomHandler.adjustToSize(xRange, yRange);
-  zoomHandler.setViewport(250, 150, 600);
+  layer.onRescale(createEventObj(container));
 
   root.appendChild(container);
-  root.appendChild(createFPSLabel());
 
   return root;
 };
