+++ .storybook/src/layers/canvas/callout-with-zoom.stories.ts
@@ -1,9 +1,10 @@
 import { scaleLinear } from 'd3-scale';
 
-import { CalloutCanvasLayer } from '../../../src/layers';
-import { Annotation } from '../../../src/interfaces';
+import { CalloutCanvasLayer } from '../../../../src/layers';
+import { ZoomPanHandler } from '../../../../src/control/ZoomPanHandler';
+import { Annotation, OnUpdateEvent } from '../../../../src/interfaces';
 
-import { createLayerContainer, createRootContainer } from '../utils';
+import { createLayerContainer, createRootContainer, createFPSLabel } from '../../utils';
 
 const annotations: Annotation[] = [
   {
@@ -25,6 +26,15 @@ const annotations: Annotation[] = [
     data: [460, 110],
   },
   {
+    title: 'Balder Fm. Top',
+    md: 1234.3,
+    tvd: 1234,
+    mdUnit: 'm',
+    depthReferencePoint: 'RKB',
+    group: 'strat-picks',
+    data: [460, 110],
+  },
+  {
     title: 'Odin Fm. Top',
     md: 1234.3,
     tvd: 1234,
@@ -32,6 +42,15 @@ const annotations: Annotation[] = [
     depthReferencePoint: 'RKB',
     group: 'strat-picks',
     data: [350, 60],
+  },
+  {
+    title: 'Loke Fm. 2.1 Top',
+    md: 1234,
+    tvd: 1234,
+    mdUnit: 'm',
+    depthReferencePoint: 'RKB',
+    group: 'strat-picks',
+    data: [40, 70],
   },
   {
     title: 'Loke Fm. 2.1 Top',
@@ -69,8 +88,11 @@ const margin = {
   left: 0,
 };
 
-const xbounds = [0, 500];
-const ybounds = [0, 500];
+const xbounds: [number, number] = [0, 500];
+const ybounds: [number, number] = [0, 500];
+
+const xRange = 500;
+const yRange = 500;
 
 const xscale = scaleLinear()
   .domain(xbounds)
@@ -82,7 +104,7 @@ const yscale = scaleLinear()
 const width = 500;
 const height = 500;
 
-export const CalloutCanvas = () => {
+export const CalloutCanvasWithZoom = () => {
   const root = createRootContainer(width);
   const container = createLayerContainer(width, height);
 
@@ -90,7 +112,23 @@ export const CalloutCanvas = () => {
   layer.onMount(createEventObj(container));
   layer.onUpdate(createEventObj(container));
 
+  const zoomHandler = new ZoomPanHandler(container, (event: OnUpdateEvent) => {
+    layer.onRescale({
+      ...event,
+      data: annotations,
+      annotations,
+      isLeftToRight: true,
+      margin,
+      scale: 0,
+    });
+  });
+
+  zoomHandler.setBounds(xbounds, ybounds);
+  zoomHandler.adjustToSize(xRange, yRange);
+  zoomHandler.setViewport(250, 150, 600);
+
   root.appendChild(container);
+  root.appendChild(createFPSLabel());
 
   return root;
 };
