+++ .storybook/src/layers/canvas/grid-layer.stories.ts
@@ -1,5 +1,7 @@
 import { scaleLinear } from 'd3-scale';
-import { GridLayer } from '../../../src/layers';
+import { GridLayer } from '../../../../src/layers';
+
+import { createRootContainer, createLayerContainer } from '../../utils';
 
 const width = 400;
 const height = 500;
@@ -8,6 +10,9 @@ const xbounds = [0, 1000];
 const ybounds = [0, 1000];
 
 export const CanvasGrid = () => {
+  const root = createRootContainer(width);
+  const container = createLayerContainer(width, height);
+
   const gridLayer = new GridLayer('grid', {
     order: 1,
     majorColor: 'black',
@@ -16,18 +21,14 @@ export const CanvasGrid = () => {
     minorWidth: 0.5,
   });
 
-  const root = document.createElement('div');
-  root.className = 'grid-container';
-  root.setAttribute('style', `height: ${height}px; width: ${width}px;background-color: #eee;position: relative;`);
-  root.setAttribute('height', `${height}`);
-  root.setAttribute('width', `${width}`);
-
-  gridLayer.onMount({ elm: root });
+  gridLayer.onMount({ elm: container });
 
   /**
    * .onUpdate(...) sets width and height of the canvas, currently .render() uses default dimensions (150, 300)
    */
-  gridLayer.onUpdate(createEventObj(root));
+  gridLayer.onUpdate(createEventObj(container));
+
+  root.appendChild(container);
 
   return root;
 };
@@ -39,10 +40,10 @@ export const CanvasGrid = () => {
 const createEventObj = (elm: any) => {
   const xscale = scaleLinear()
     .domain(xbounds)
-    .range([0, 500]);
+    .range([0, width]);
   const yscale = scaleLinear()
     .domain(ybounds)
-    .range([0, 500]);
+    .range([0, width]);
 
   return {
     xScale: xscale.copy(),
