+++ .storybook/src/layers/canvas/index.stories.ts
@@ -1,0 +1,8 @@
+export * from './callout-with-zoom.stories';
+export * from './multiple-layers.stories';
+export * from './callout-canvas.layer.stories';
+export * from './grid-layer.stories';
+
+export default {
+  title: 'Canvas',
+};
