+++ .storybook/src/layers/svg/index.stories.ts
@@ -1,0 +1,5 @@
+export * from './wellborepath-layer.stories';
+
+export default {
+  title: 'SVG',
+};
