+++ .storybook/src/layers/webgl/index.stories.ts
@@ -1,0 +1,6 @@
+export * from './HoleSizeLayer.stories';
+export * from './geomodel-layer.stories';
+
+export default {
+  title: 'PIXI JS WebGL Layer',
+};
